ci(travis): run semantic_release on after_success
semantic_release is not a deployment so it make more sense to run it on
after_success and not on deploy. It handles by it self the cases where
it must not run, as on a pull request, which is why it was on deploy as
deploy is skipped on pull requests.